reactor(apix): endpoint differs from query
Using objects for the various endpoints allows us to remove a lot of redundant
code and also remove the non-sense of having to construct queries that are also
endpoints, forcing an additional invoke operator `()` call.

Please note that last commit includes changes to OpenMeteo that should have been
committed in this one instead.